# docs/external-event-api.openapi.yaml
openapi: 3.1.0
info:
  title: Dance Billboard External Event API
  version: 1.3.0
  description: 외부 파트너가 소유한 Dance Billboard 일정을 등록·수정·삭제하는 서버 간 API
servers:
  - url: https://swingenjoy.com
paths:
  /api/external/v1/events:
    post:
      summary: 외부 일정 등록
      operationId: createExternalEvent
      security:
        - partnerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExternalEventRequest'
      responses:
        '201':
          description: 최초 등록 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '200':
          description: 같은 external_id로 이미 등록된 일정
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DuplicateResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/external/v1/regular-socials:
    post:
      summary: 정규 소셜 반복 규칙 등록 또는 동일 ID 갱신
      operationId: createRegularSocialRule
      security:
        - partnerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegularSocialRuleRequest'
      responses:
        '201':
          description: 규칙 등록 성공
        '200':
          description: 동일 규칙 갱신 또는 테스트 검증 성공
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/external/v1/regular-socials/{external_id}:
    parameters:
      - name: external_id
        in: path
        required: true
        schema: { type: string, maxLength: 160 }
    put:
      summary: 정규 소셜 반복 규칙 수정
      operationId: updateRegularSocialRule
      security:
        - partnerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegularSocialRuleRequest'
      responses:
        '200': { description: 수정 성공 }
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: 정규 소셜 반복 규칙과 연결 예외 삭제
      operationId: deleteRegularSocialRule
      security:
        - partnerApiKey: []
      responses:
        '200': { description: 삭제 성공 }
        '404':
          $ref: '#/components/responses/NotFound'
  /api/external/v1/regular-socials/{external_id}/exceptions:
    parameters:
      - name: external_id
        in: path
        required: true
        description: 반복 규칙 external_id
        schema: { type: string, maxLength: 160 }
    post:
      summary: 날짜별 휴무 또는 변경 예외 등록
      operationId: createRegularSocialException
      security:
        - partnerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegularSocialExceptionRequest'
      responses:
        '201': { description: 예외 등록 성공 }
        '200': { description: 동일 예외 갱신 또는 테스트 검증 성공 }
        '404':
          $ref: '#/components/responses/NotFound'
  /api/external/v1/regular-socials/{external_id}/exceptions/{exception_id}:
    parameters:
      - name: external_id
        in: path
        required: true
        description: 반복 규칙 external_id
        schema: { type: string, maxLength: 160 }
      - name: exception_id
        in: path
        required: true
        schema: { type: string, maxLength: 160 }
    put:
      summary: 날짜별 예외 수정
      operationId: updateRegularSocialException
      security:
        - partnerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegularSocialExceptionRequest'
      responses:
        '200': { description: 수정 성공 }
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: 날짜별 예외 삭제
      operationId: deleteRegularSocialException
      security:
        - partnerApiKey: []
      responses:
        '200': { description: 삭제 성공 }
        '404':
          $ref: '#/components/responses/NotFound'
  /api/external/v1/images:
    post:
      summary: 외부 일정 이미지 업로드
      operationId: uploadExternalEventImage
      security:
        - partnerApiKey: []
      requestBody:
        required: true
        content:
          image/jpeg:
            schema:
              type: string
              contentMediaType: image/jpeg
              contentEncoding: binary
          image/png:
            schema:
              type: string
              contentMediaType: image/png
              contentEncoding: binary
          image/webp:
            schema:
              type: string
              contentMediaType: image/webp
              contentEncoding: binary
          image/avif:
            schema:
              type: string
              contentMediaType: image/avif
              contentEncoding: binary
      responses:
        '201':
          description: 이미지 검사·변환·저장 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageUploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/external/v1/events/{external_id}:
    parameters:
      - name: external_id
        in: path
        required: true
        schema:
          type: string
          maxLength: 160
    put:
      summary: 동일 파트너가 등록한 일정 전체 수정
      operationId: updateExternalEvent
      security:
        - partnerApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExternalEventRequest'
      responses:
        '200':
          description: 수정 성공
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
    delete:
      summary: 동일 파트너가 등록한 일정 삭제
      operationId: deleteExternalEvent
      security:
        - partnerApiKey: []
      responses:
        '200':
          description: 삭제 성공
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
components:
  securitySchemes:
    partnerApiKey:
      type: http
      scheme: bearer
      bearerFormat: SwingEnjoyPartnerKey
  schemas:
    RegularSocialRuleRequest:
      type: object
      additionalProperties: false
      required: [external_id, title, weekday, location, source_url]
      properties:
        external_id: { type: string, maxLength: 160 }
        title: { type: string, maxLength: 255 }
        weekday:
          type: integer
          minimum: 0
          maximum: 6
          description: 일요일 0, 토요일 6
        time: { type: string, maxLength: 255 }
        location: { type: string, maxLength: 255 }
        venue_name: { type: string, maxLength: 255 }
        source_url: { type: string, format: uri, pattern: '^https://' }
        source_id: { type: string, maxLength: 160 }
        valid_from: { type: string, format: date }
        valid_until: { type: string, format: date }
        active: { type: boolean, default: true }
    RegularSocialExceptionRequest:
      type: object
      additionalProperties: false
      required: [external_id, date, type]
      properties:
        external_id: { type: string, maxLength: 160 }
        date: { type: string, format: date }
        type:
          type: string
          enum: [closure, override]
        title: { type: string, maxLength: 255 }
        time: { type: string, maxLength: 255 }
        location: { type: string, maxLength: 255 }
        venue_name: { type: string, maxLength: 255 }
        dj_name: { type: string, maxLength: 255 }
        source_url: { type: string, format: uri, pattern: '^https://' }
        description: { type: string, maxLength: 20000 }
    ExternalEventRequest:
      type: object
      additionalProperties: false
      required:
        - external_id
        - title
        - source_url
        - event_dates
      allOf:
        - if:
            properties:
              category:
                enum: [event, class, club]
            required: [category]
          then:
            required: [image_mode, image_url]
      properties:
        external_id:
          type: string
          maxLength: 160
        title:
          type: string
          maxLength: 255
        event_dates:
          type: array
          minItems: 1
          maxItems: 366
          uniqueItems: true
          items:
            type: string
            format: date
          description: 실제 노출할 날짜. 단일 일정은 1개, 개별 날짜 일정은 여러 개를 전송.
        time:
          type: string
          maxLength: 120
        location:
          type: string
          maxLength: 255
        address:
          type: string
          maxLength: 255
          description: 선택값. 상세 지도는 부정확한 입력의 검색 첫 결과를 표시할 수 있으므로 확인된 도로명주소 권장
        address_detail:
          type: string
          maxLength: 255
          description: 층·호수 등 지도 검색에서 분리할 상세 위치
        postal_code:
          type: string
          pattern: '^\d{5}$'
        address_source:
          type: string
          enum: [daum_postcode, kakao_map, naver_map, google_map, road_address_api]
          description: 선택값. 주소를 확인한 서비스 기록
        location_link:
          type: string
          format: uri
          pattern: '^https://'
        description:
          type: string
          maxLength: 20000
        category:
          type: string
          enum: [social, event, class, club]
          description: 파트너 기본값이 설정된 경우 생략 가능
        genre:
          type: string
          maxLength: 64
          description: 선택한 category에 허용된 값만 사용합니다. event만 쉼표로 구분한 복수 장르를 허용하며 파티+대회는 동시 사용 불가, 기타는 단독만 가능합니다. social/class/club은 1개만 사용합니다.
        source_url:
          type: string
          format: uri
          pattern: '^https://'
        link_name1:
          type: string
          maxLength: 120
        image_url:
          type: string
          format: uri
          pattern: '^https://.*\\.(avif|jpg|jpeg|png|webp)([?#].*)?$'
        image_mode:
          type: string
          enum: [upload, url]
          description: upload은 이미지 업로드 API 반환 URL, url은 외부 공개 이미지 URL
        venue_name:
          type: string
          maxLength: 255
    SuccessResponse:
      type: object
      required: [ok, duplicate, event_id, event]
      properties:
        ok:
          type: boolean
          const: true
        duplicate:
          type: boolean
          const: false
        event_id:
          type: string
        event:
          type: object
    DuplicateResponse:
      type: object
      required: [ok, duplicate, event_id]
      properties:
        ok:
          type: boolean
          const: true
        duplicate:
          type: boolean
          const: true
        event_id:
          type: string
    ImageUploadResponse:
      type: object
      required: [ok, image_url, variants, content_type, bytes]
      properties:
        ok:
          type: boolean
          const: true
        image_url:
          type: string
          format: uri
        variants:
          type: object
          required: [image_micro, image_thumbnail, image_medium, image_full]
          properties:
            image_micro: { type: string, format: uri }
            image_thumbnail: { type: string, format: uri }
            image_medium: { type: string, format: uri }
            image_full: { type: string, format: uri }
        content_type:
          type: string
          const: image/webp
        bytes:
          type: object
          additionalProperties:
            type: integer
            minimum: 1
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
  responses:
    BadRequest:
      description: 요청값 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: API Key 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 해당 파트너 소유 일정 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PayloadTooLarge:
      description: 요청 본문 크기 초과
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnsupportedMediaType:
      description: 지원하지 않는 이미지 Content-Type
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: 호출량 제한 초과
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'