# server/cafe24/app.js
import dotenv from 'dotenv';
import express from 'express';
import path from 'node:path';
import { existsSync, readFileSync } from 'node:fs';
import { createCafe24FunctionHandler } from './cafe24-function-adapter.js';
import { elapsedMs, nowMs, perfLog, perfShouldLog, summarizeApiRequest } from './perf-log.js';
import {
  createCafe24Event,
  deleteCafe24Event,
  getCafe24Event,
  listCafe24Events,
  updateCafe24Event,
} from './events-api.js';
import {
  createExternalEvent,
  createExternalPartner,
  autoIncreaseExternalTestLimit,
  deleteExternalEvent,
  listExternalAdminAuditLogs,
  listExternalPartners,
  listMyExternalPartners,
  listExternalPartnerRequests,
  requestExternalPartnerAccess,
  listExternalRequestLogs,
  rotateExternalPartnerKey,
  searchExternalPartnerUsers,
  updateExternalEvent,
  updateExternalPartnerStatus,
  uploadExternalEventImage,
} from './external-events-api.js';
import {
  createRegularSocialRule,
  deleteRegularSocialException,
  deleteRegularSocialRule,
  updateRegularSocialRule,
  upsertRegularSocialException,
} from './external-regular-socials-api.js';
import { authProviders, devLogin, googleLoginCallback, googleLoginStart, kakaoLogin, logout, me } from './auth-api.js';
import {
  listClientReloadDiagnostics,
  listServerVersionDiagnostics,
  recordClientReloadDiagnostic,
} from './diagnostics-api.js';
import { uploadEventImage } from './uploads-api.js';
import { reconcileRegularSocials } from './regular-social-reconciler.js';
import { eventStats, recordAnalytics, siteStats } from './stats-api.js';
import {
  dailyDigestCron,
  listUserNotifications,
  markUserNotificationsRead,
  notificationQueueCron,
  processNotificationQueue,
  sendBoardCommentNotification,
  sendDailyDigestToAdmins,
  sendPushNotification,
} from './push-api.js';
import {
  cafe24IngestorV3Candidates,
  cafe24IngestorV3RegisterBlocked,
  cafe24IngestorV3ReviewCandidate,
} from './ingestor-v3-api.js';
import {
  callRpc,
  deleteRecords,
  insertRecords,
  listGenericFiles,
  queryRecords,
  removeGenericFiles,
  updateRecords,
  uploadGenericFile,
  upsertRecords,
} from './generic-data-api.js';
import {
  cafe24BillboardManifest,
  cafe24DeleteEventFunction,
  cafe24DeleteInvitation,
  cafe24DeleteSocialItem,
  cafe24IngestorRegisterEvent,
  cafe24Invitations,
  cafe24OneDayRecruitLogo,
  cafe24ScrapedEvents,
  cafe24UnavailableFunction,
  cafe24ValidateInvitation,
} from './function-api.js';

dotenv.config({ path: path.resolve(process.cwd(), '.env') });
dotenv.config({ path: path.resolve(process.cwd(), '.env.local') });

const app = express();
const distDir = path.resolve(process.cwd(), process.env.CAFE24_DIST_DIR || 'dist');
const uploadsDir = path.resolve(process.cwd(), process.env.CAFE24_UPLOADS_DIR || 'uploads');
const uploadsFallbackOrigin = String(process.env.CAFE24_UPLOADS_FALLBACK_ORIGIN || '').replace(/\/+$/g, '');
const rawBody = express.raw({
  type: '*/*',
  limit: process.env.CAFE24_BODY_LIMIT || '50mb',
});

function isSwingFloorCouncilRoute(reqPath = '') {
  return reqPath === '/swing-floor-council' || reqPath === '/swing-floor-council/';
}

function buildSwingFloorCouncilPreviewHtml(html = '') {
  const shareImageUrl = 'https://swingenjoy.com/swing-floor-council-share-card.png?v=20260626-1';
  const title = '스윙 플로어 협의체 발기인 모집';
  const description = '스윙 플로어를 지키기 위한 협의체 발기인 모집 초안입니다.';
  const metadata = `
  <title>${title}</title>
  <meta name="description" content="${description}" />
  <meta property="og:type" content="website" />
  <meta property="og:url" content="https://swingenjoy.com/swing-floor-council" />
  <meta property="og:title" content="${title}" />
  <meta property="og:description" content="${description}" />
  <meta property="og:image" content="${shareImageUrl}" />
  <meta property="og:image:width" content="1200" />
  <meta property="og:image:height" content="630" />
  <meta property="twitter:card" content="summary_large_image" />
  <meta property="twitter:title" content="${title}" />
  <meta property="twitter:description" content="${description}" />
  <meta property="twitter:image" content="${shareImageUrl}" />
  <meta name="robots" content="noindex,nofollow,noarchive" />`;

  return html
    .replace(/<meta\s+[^>]*(?:property|name)=["'](?:og:[^"']+|twitter:[^"']+)["'][^>]*>\s*/gi, '')
    .replace(/<meta\s+[^>]*name=["']description["'][^>]*>\s*/gi, '')
    .replace(/<meta\s+[^>]*name=["']apple-mobile-web-app-title["'][^>]*>\s*/gi, '')
    .replace(/<link\s+[^>]*rel=["']manifest["'][^>]*>\s*/gi, '')
    .replace(/<title>[\s\S]*?<\/title>/i, '')
    .replace('</head>', `${metadata}\n</head>`);
}

function isExternalEventApiGuideRoute(reqPath = '') {
  return reqPath === '/external-event-api' || reqPath === '/external-event-api/';
}

function buildExternalEventApiGuideHtml(html = '') {
  const title = 'Dance Billboard 외부 일정 연동 API';
  const description = '외부 사이트에서 일정을 안전하게 등록·수정·삭제하는 API 연동 매뉴얼입니다.';
  const metadata = `
  <title>${title}</title>
  <meta name="description" content="${description}" />
  <link rel="canonical" href="https://swingenjoy.com/external-event-api" />
  <meta property="og:type" content="website" />
  <meta property="og:url" content="https://swingenjoy.com/external-event-api" />
  <meta property="og:title" content="${title}" />
  <meta property="og:description" content="${description}" />
  <meta property="og:image" content="https://swingenjoy.com/kakao-share-card.png" />
  <meta property="twitter:card" content="summary_large_image" />`;
  return html
    .replace(/<meta\s+[^>]*(?:property|name)=["'](?:og:[^"']+|twitter:[^"']+)["'][^>]*>\s*/gi, '')
    .replace(/<meta\s+[^>]*name=["']description["'][^>]*>\s*/gi, '')
    .replace(/<link\s+[^>]*rel=["']canonical["'][^>]*>\s*/gi, '')
    .replace(/<title>[\s\S]*?<\/title>/i, '')
    .replace('</head>', `${metadata}\n</head>`);
}
const urlencodedBody = express.urlencoded({
  extended: false,
  limit: process.env.CAFE24_SHARE_TARGET_BODY_LIMIT || '128kb',
});
const jsonBody = express.json({
  limit: process.env.CAFE24_BODY_LIMIT || '50mb',
});
const externalEventJsonBody = express.json({ limit: '256kb' });
const externalImageBody = express.raw({
  type: ['image/jpeg', 'image/png', 'image/webp', 'image/avif'],
  limit: process.env.EXTERNAL_IMAGE_MAX_BYTES || '32mb',
});
const cafe24FunctionHandler = createCafe24FunctionHandler();

app.disable('x-powered-by');
app.set('trust proxy', true);

app.use((_req, res, next) => {
  res.setHeader('X-Content-Type-Options', 'nosniff');
  res.setHeader('X-XSS-Protection', '1; mode=block');
  res.setHeader('Referrer-Policy', 'strict-origin-when-cross-origin');
  res.setHeader('Permissions-Policy', 'geolocation=(), microphone=(), camera=()');
  res.setHeader('X-DNS-Prefetch-Control', 'on');
  res.setHeader('Content-Security-Policy', "frame-ancestors 'self' https://xn--9m1bu8iv0ao6ao5fsta.com");
  next();
});

app.use((req, res, next) => {
  if (!String(req.path || '').startsWith('/api/')) {
    next();
    return;
  }

  const startedAt = nowMs();
  res.on('finish', () => {
    const pathName = req.path || req.originalUrl || '';
    const ms = elapsedMs(startedAt);
    const always =
      pathName === '/api/auth/me' ||
      pathName === '/api/analytics/session' ||
      pathName.startsWith('/api/auth/google/') ||
      pathName.startsWith('/api/cafe24-rpc/get_user_interactions') ||
      pathName.startsWith('/api/cafe24-rpc/increment_item_views') ||
      pathName.startsWith('/api/cafe24-data/board_');

    if (!perfShouldLog(ms, always)) return;

    perfLog('http', {
      method: req.method,
      path: pathName,
      status: res.statusCode,
      ms,
      ...summarizeApiRequest(req),
    }, ms >= 1000 || res.statusCode >= 500 ? 'warn' : 'log');
  });
  next();
});

app.get('/__health', (_req, res) => {
  res.json({
    ok: true,
    dist: existsSync(path.join(distDir, 'index.html')),
    uploads: uploadsDir,
    uploadsFallbackOrigin: uploadsFallbackOrigin || null,
    functions: process.env.CAFE24_FUNCTIONS_DIR || 'dist-cafe24/functions',
  });
});

async function proxyMissingUpload(req, res, next) {
  if (!uploadsFallbackOrigin) {
    next();
    return;
  }

  try {
    const uploadPath = req.originalUrl.startsWith('/uploads/')
      ? req.originalUrl
      : `/uploads${req.url}`;
    const remoteUrl = new URL(uploadPath, uploadsFallbackOrigin);
    const response = await fetch(remoteUrl, {
      headers: {
        Accept: req.get('accept') || 'image/avif,image/webp,image/apng,image/*,*/*;q=0.8',
        'User-Agent': 'SwingEnjoyLocalProdDb/1.0',
      },
    });

    const contentType = response.headers.get('content-type') || '';
    if (!response.ok || contentType.toLowerCase().includes('text/html')) {
      next();
      return;
    }

    const contentLength = response.headers.get('content-length');
    if (contentType) res.setHeader('Content-Type', contentType);
    if (contentLength) res.setHeader('Content-Length', contentLength);
    res.setHeader('Cache-Control', 'public, max-age=300');
    res.setHeader('X-Uploads-Fallback-Origin', uploadsFallbackOrigin);
    const body = Buffer.from(await response.arrayBuffer());
    res.status(response.status).send(body);
  } catch (error) {
    console.warn('[cafe24:uploads-fallback]', error?.message || error);
    next();
  }
}

function alias(functionName) {
  return (req, _res, next) => {
    req.cafe24FunctionName = functionName;
    next();
  };
}

function jsonRoute(handler) {
  return async (req, res) => {
    try {
      await handler(req, res);
    } catch (error) {
      const statusCode = Number(error?.statusCode || 500);
      if (statusCode >= 500) {
        console.error('[cafe24:api]', error);
      } else {
        console.warn('[cafe24:api]', statusCode, error?.message || 'Request rejected');
      }
      res.status(statusCode).json({
        error: 'Cafe24 API Error',
        code: statusCode < 500 ? error?.code : undefined,
        message: error?.message || 'Unknown Cafe24 API error',
      });
    }
  };
}

app.get('/api/events', jsonRoute(listCafe24Events));
app.get('/api/events/:id', jsonRoute(getCafe24Event));
app.post('/api/events', jsonBody, jsonRoute(createCafe24Event));
app.post('/api/events/:id', jsonBody, jsonRoute(updateCafe24Event));
app.post('/api/events/:id/delete', jsonBody, jsonRoute(deleteCafe24Event));
app.put('/api/events/:id', jsonBody, jsonRoute(updateCafe24Event));
app.patch('/api/events/:id', jsonBody, jsonRoute(updateCafe24Event));
app.delete('/api/events/:id', jsonRoute(deleteCafe24Event));
app.post('/api/external/v1/events', externalEventJsonBody, jsonRoute(createExternalEvent));
app.put('/api/external/v1/events/:externalId', externalEventJsonBody, jsonRoute(updateExternalEvent));
app.delete('/api/external/v1/events/:externalId', jsonRoute(deleteExternalEvent));
app.post('/api/external/v1/regular-socials', externalEventJsonBody, jsonRoute(createRegularSocialRule));
app.put('/api/external/v1/regular-socials/:externalId', externalEventJsonBody, jsonRoute(updateRegularSocialRule));
app.delete('/api/external/v1/regular-socials/:externalId', jsonRoute(deleteRegularSocialRule));
app.post(
  '/api/external/v1/regular-socials/:externalId/exceptions',
  externalEventJsonBody,
  jsonRoute(upsertRegularSocialException),
);
app.put(
  '/api/external/v1/regular-socials/:externalId/exceptions/:exceptionId',
  externalEventJsonBody,
  jsonRoute(upsertRegularSocialException),
);
app.delete(
  '/api/external/v1/regular-socials/:externalId/exceptions/:exceptionId',
  jsonRoute(deleteRegularSocialException),
);
app.post('/api/external/v1/images', externalImageBody, jsonRoute(uploadExternalEventImage));
app.get('/api/admin/external-partners', jsonRoute(listExternalPartners));
app.get('/api/admin/external-partner-requests', jsonRoute(listExternalPartnerRequests));
app.post('/api/external/partner-requests', jsonBody, jsonRoute(requestExternalPartnerAccess));
app.get('/api/external/my-partners', jsonRoute(listMyExternalPartners));
app.post('/api/external/my-partners/:partnerId/auto-test-limit', jsonBody, jsonRoute(autoIncreaseExternalTestLimit));
app.post('/api/admin/external-partners', jsonBody, jsonRoute(createExternalPartner));
app.patch('/api/admin/external-partners/:partnerId', jsonBody, jsonRoute(updateExternalPartnerStatus));
app.post('/api/admin/external-partners/:partnerId/rotate-key', jsonBody, jsonRoute(rotateExternalPartnerKey));
app.get('/api/admin/external-partner-users', jsonRoute(searchExternalPartnerUsers));
app.get('/api/admin/external-request-logs', jsonRoute(listExternalRequestLogs));
app.get('/api/admin/external-admin-audit-logs', jsonRoute(listExternalAdminAuditLogs));
app.use('/api/external/v1', (error, req, res, next) => {
  if (error?.type === 'entity.too.large') {
    const isImageUpload = req.path === '/images' || req.originalUrl?.startsWith('/api/external/v1/images');
    res.status(413).json({
      error: 'External Event API Error',
      code: 'payload_too_large',
      message: isImageUpload
        ? '이미지 파일은 32MB를 초과할 수 없습니다.'
        : '일정 JSON 본문은 256KB를 초과할 수 없습니다.',
    });
    return;
  }
  if (error instanceof SyntaxError && error?.status === 400) {
    res.status(400).json({
      error: 'External Event API Error',
      code: 'invalid_json',
      message: '올바른 JSON 본문이 필요합니다.',
    });
    return;
  }
  next(error);
});

app.get('/api/auth/me', jsonRoute(me));
app.get('/api/auth/providers', jsonRoute(authProviders));
app.post('/api/auth/dev-login', jsonBody, jsonRoute(devLogin));
app.get('/api/auth/dev-login', jsonRoute(devLogin));
app.post('/api/auth/logout', jsonRoute(logout));
app.post('/api/kakao-login', jsonBody, jsonRoute(kakaoLogin));
app.post('/api/auth/kakao', jsonBody, jsonRoute(kakaoLogin));
app.get('/api/auth/google/start', jsonRoute(googleLoginStart));
app.get('/api/auth/google/callback', jsonRoute(googleLoginCallback));

app.post('/api/uploads/events', jsonBody, jsonRoute(uploadEventImage));

app.post('/api/cafe24-data/:table/query', jsonBody, jsonRoute(queryRecords));
app.post('/api/cafe24-data/:table/insert', jsonBody, jsonRoute(insertRecords));
app.post('/api/cafe24-data/:table/update', jsonBody, jsonRoute(updateRecords));
app.post('/api/cafe24-data/:table/upsert', jsonBody, jsonRoute(upsertRecords));
app.post('/api/cafe24-data/:table/delete', jsonBody, jsonRoute(deleteRecords));
app.post('/api/cafe24-rpc/:name', jsonBody, jsonRoute(callRpc));
app.post('/api/cafe24-storage/:bucket/upload', jsonBody, jsonRoute(uploadGenericFile));
app.get('/api/cafe24-storage/:bucket/list', jsonRoute(listGenericFiles));
app.delete('/api/cafe24-storage/:bucket/remove', jsonBody, jsonRoute(removeGenericFiles));
app.post('/api/cafe24-storage/:bucket/remove', jsonBody, jsonRoute(removeGenericFiles));

app.get('/api/stats/events', jsonRoute(eventStats));
app.get('/api/stats/site', jsonRoute(siteStats));
app.post('/api/analytics/session', jsonBody, jsonRoute(recordAnalytics));
app.post('/api/diagnostics/reload', jsonBody, jsonRoute(recordClientReloadDiagnostic));
app.get('/api/diagnostics/reloads', jsonRoute(listClientReloadDiagnostics));
app.get('/api/diagnostics/server-versions', jsonRoute(listServerVersionDiagnostics));
app.post('/api/send-push-notification', jsonBody, jsonRoute(sendPushNotification));
app.post('/api/board/comment-notification', jsonBody, jsonRoute(sendBoardCommentNotification));
app.get('/api/notifications', jsonRoute(listUserNotifications));
app.post('/api/notifications/read', jsonBody, jsonRoute(markUserNotificationsRead));
app.post('/api/admin/push/send-daily-digest-test', jsonBody, jsonRoute(sendDailyDigestToAdmins));
app.post('/api/admin/push/process-notification-queue', jsonBody, jsonRoute(processNotificationQueue));
app.all('/api/__cron/daily-digest', jsonBody, jsonRoute(dailyDigestCron));
app.all('/api/__cron/notification-queue', jsonBody, jsonRoute(notificationQueueCron));
app.all('/api/__cron/regular-socials', jsonBody, jsonRoute(reconcileRegularSocials));

app.all('/api/scraped-events', jsonBody, jsonRoute(cafe24ScrapedEvents));
app.post('/api/ingestor-register-event', jsonBody, jsonRoute(cafe24IngestorRegisterEvent));
app.all('/api/ingestor-v3/candidates', jsonBody, jsonRoute(cafe24IngestorV3Candidates));
app.post('/api/ingestor-v3/candidates/:id/review', jsonBody, jsonRoute(cafe24IngestorV3ReviewCandidate));
app.patch('/api/ingestor-v3/candidates/:id/review', jsonBody, jsonRoute(cafe24IngestorV3ReviewCandidate));
app.post('/api/ingestor-v3/candidates/:id/register', jsonBody, jsonRoute(cafe24IngestorV3RegisterBlocked));
app.post('/api/delete-event', jsonBody, jsonRoute(cafe24DeleteEventFunction));
app.post('/api/delete-social-item', jsonBody, jsonRoute(cafe24DeleteSocialItem));
app.post('/api/oneday-recruit-logo', jsonBody, jsonRoute(cafe24OneDayRecruitLogo));
app.all('/api/fetch-og-image', rawBody, alias('fetch-og-image'), cafe24FunctionHandler);
app.all('/api/tango-scene-map', rawBody, alias('tango-scene-map'), cafe24FunctionHandler);
app.delete('/api/invitations/delete', jsonBody, jsonRoute(cafe24DeleteInvitation));
app.post('/api/invitations/delete', jsonBody, jsonRoute(cafe24DeleteInvitation));

app.all('/api/invitations', jsonBody, jsonRoute(cafe24Invitations));
app.post('/api/invitations/validate', jsonBody, jsonRoute(cafe24ValidateInvitation));
app.delete('/api/invitations/:id', jsonBody, jsonRoute(cafe24DeleteInvitation));
app.get('/api/billboard-manifest', jsonRoute(cafe24BillboardManifest));

function compactShareTargetText(value) {
  return String(value || '').replace(/\s+/g, ' ').trim();
}

function trimShareTargetText(value) {
  return String(value || '').trim();
}

function extractShareTargetUrl(value) {
  const match = compactShareTargetText(value).match(/https?:\/\/[^\s"'<>]+/i);
  return match ? match[0].replace(/[),.?!\]]+$/, '') : '';
}

function stripShareTargetUrl(value, url) {
  let text = compactShareTargetText(value);
  if (url) text = text.replace(url, ' ');
  return compactShareTargetText(text.replace(/https?:\/\/[^\s"'<>]+/gi, ' '));
}

function buildShareTargetDraft(body = {}) {
  const url = compactShareTargetText(body.url || body.add || body.link)
    || extractShareTargetUrl(body.text)
    || extractShareTargetUrl(body.title);
  const title = compactShareTargetText(body.title);
  const description = trimShareTargetText(stripShareTargetUrl(body.text, url));

  return {
    form: {
      url,
      title,
      description,
      authorName: '',
      thumbnailUrl: '',
      archiveBucket: 'reference',
      collectionName: '',
      tags: '',
      danceGenre: '',
      sourceContext: 'Android 공유',
      publishedAt: '',
    },
    savedAt: Date.now(),
    source: 'android-share',
  };
}

function sendShareTargetBootstrap(req, res) {
  const draft = buildShareTargetDraft(req.body || {});
  const draftJson = JSON.stringify(draft).replace(/</g, '\\u003c');
  res
    .status(200)
    .type('html')
    .send(`<!doctype html>
<html lang="ko">
<head>
  <meta charset="utf-8">
  <meta name="viewport" content="width=device-width, initial-scale=1">
  <title>공유 저장 중</title>
</head>
<body>
  <script>
    (function () {
      try {
        var draft = ${draftJson};
        if (draft && draft.form && (draft.form.url || draft.form.title || draft.form.description)) {
          localStorage.setItem('swingenjoy:media-archive-pending-draft', JSON.stringify(draft));
        }
      } catch (error) {}
      location.replace('/forum/media');
    }());
  </script>
</body>
</html>`);
}

app.post('/forum/media/share', urlencodedBody, sendShareTargetBootstrap);

app.use('/uploads', express.static(uploadsDir, {
  etag: true,
  maxAge: '30d',
}));
app.use('/uploads', proxyMissingUpload);
app.use('/uploads', (_req, res) => {
  res.setHeader('Cache-Control', 'no-store');
  res.status(404).type('text/plain').send('Upload not found');
});

app.all(/^\/api\/([^/?#]+)(?:\/.*)?$/, rawBody, async (req, res, next) => {
  try {
    req.params.name = req.params[0];
    await cafe24UnavailableFunction(req, res);
  } catch (error) {
    next(error);
  }
});

app.use(express.static(distDir, {
  index: false,
  etag: true,
  setHeaders(res, filePath) {
    if (filePath.includes(`${path.sep}assets${path.sep}`)) {
      res.setHeader('Cache-Control', 'public, max-age=31536000, immutable');
      return;
    }

    if (filePath.endsWith('service-worker.js') || filePath.endsWith('version.json')) {
      res.setHeader('Cache-Control', 'no-cache, no-store, must-revalidate');
      return;
    }

    if (filePath.endsWith('manifest.json')) {
      res.setHeader('Cache-Control', 'public, max-age=300, stale-while-revalidate=300');
      res.setHeader('Content-Type', 'application/manifest+json; charset=utf-8');
    }
  },
}));

app.use((req, res, next) => {
  if (!['GET', 'HEAD'].includes(req.method)) {
    next();
    return;
  }

  const acceptsHtml = req.accepts(['html', 'json']) === 'html';
  const isPwaShareTarget = req.path === '/forum/media/share';
  if (!acceptsHtml && !isPwaShareTarget) {
    next();
    return;
  }

  const indexFile = path.join(distDir, 'index.html');
  if (!existsSync(indexFile)) {
    res.status(500).send('Cafe24 dist/index.html not found. Run npm run build:cafe24 first.');
    return;
  }

  if (isSwingFloorCouncilRoute(req.path)) {
    const html = readFileSync(indexFile, 'utf8');
    res.setHeader('Cache-Control', 'no-cache, no-store, must-revalidate');
    res.type('html').send(buildSwingFloorCouncilPreviewHtml(html));
    return;
  }
  if (isExternalEventApiGuideRoute(req.path)) {
    const html = readFileSync(indexFile, 'utf8');
    res.setHeader('Cache-Control', 'no-cache, no-store, must-revalidate');
    res.type('html').send(buildExternalEventApiGuideHtml(html));
    return;
  }

  res.setHeader('Cache-Control', 'no-cache');
  res.sendFile(indexFile);
});

export default app;